ENH Cost basis now takes into account sales
The standard definition of the cost basis, as implemented in Catalyst,
is the weighted average of the prices of the buys into a position, with
the weights being the amount of each transactions. The problem is that
in case of sale, the cost basis as computed above does not reflect any
more the price paid for the current position and is therefore useless
to compute the profitability of the position.

The present fix computes the cost basis using both buys and sell.